+++ public/index.html
@@ -46,7 +46,7 @@
 
       <!-- Контент -->
     </main>
-  <script src = "../src/main.js"> </script>
+  <script type="module" src = "../src/main.js"> </script>
   
   </body>
 </html>
