+++ src/components/loadbutton.js
@@ -1,0 +1,5 @@
+const createLoadMoreBtn = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
+
+export {createLoadMoreBtn};
