+++ webpack.config.js
@@ -1,0 +1,12 @@
+const path = require(`path`);
+
+module.exports = {
+  mode: `development`, // Режим сборки
+  entry: `./src/main.js`, // Точка входа приложения
+  output: { // Настройка выходного файла
+    filename: `bundle.js`,
+    path: path.join(_dirname, `public`)
+  }
+  devtool: `source-map` // Подключаем sourcemaps
+
+};
